+++ changelog.md
@@ -1,3 +1,9 @@
+## 1.0.2
+
+- Add [next-css](https://flaviocopes.com/nextjs-styling-components-css/)
+- MockData with test setting
+- react-table v6
+
 ## 1.0.1
 
 - refactor apps